fix: app state to metadata conversion logic
